Move uncommon award tests from packages to awardng plugins
