+++ openprocurement/auctions/core/plugins/awarding/v1/tests/blanks/award_blanks.py
@@ -1,0 +1,754 @@
+# -*- coding: utf-8 -*-
+from openprocurement.auctions.flash.tests.base import test_organization
+
+# AuctionAwardResourceTest
+
+
+def create_auction_award_invalid(self):
+    request_path = '/auctions/{}/awards'.format(self.auction_id)
+    response = self.app.post(request_path, 'data', status=415)
+    self.assertEqual(response.status, '415 Unsupported Media Type')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description':
+            u"Content-Type header should be one of ['application/json']", u'location': u'header', u'name': u'Content-Type'}
+    ])
+
+    response = self.app.post(
+        request_path, 'data', content_type='application/json', status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Expecting value: line 1 column 1 (char 0)',
+            u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, 'data', status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Data not available',
+            u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(
+        request_path, {'not_data': {}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Data not available',
+            u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {
+                                  'invalid_field': 'invalid_value'}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Rogue field', u'location':
+            u'body', u'name': u'invalid_field'}
+    ])
+
+    response = self.app.post_json(request_path, {
+                                  'data': {'suppliers': [{'identifier': 'invalid_value'}]}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': {u'identifier': [
+            u'Please use a mapping for this field or Identifier instance instead of unicode.']}, u'location': u'body', u'name': u'suppliers'}
+    ])
+
+    response = self.app.post_json(request_path, {
+                                  'data': {'suppliers': [{'identifier': {'id': 0}}]}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [{u'contactPoint': [u'This field is required.'], u'identifier': {u'scheme': [u'This field is required.']}, u'name': [u'This field is required.'], u'address': [u'This field is required.']}], u'location': u'body', u'name': u'suppliers'},
+        {u'description': [u'This field is required.'], u'location': u'body', u'name': u'bid_id'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {'suppliers': [
+                                  {'name': 'name', 'identifier': {'uri': 'invalid_value'}}]}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [{u'contactPoint': [u'This field is required.'], u'identifier': {u'scheme': [u'This field is required.'], u'id': [u'This field is required.'], u'uri': [u'Not a well formed URL.']}, u'address': [u'This field is required.']}], u'location': u'body', u'name': u'suppliers'},
+        {u'description': [u'This field is required.'], u'location': u'body', u'name': u'bid_id'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {
+        'suppliers': [test_organization],
+        'status': 'pending',
+        'bid_id': self.initial_bids[0]['id'],
+        'lotID': '0' * 32
+    }}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [u'lotID should be one of lots'], u'location': u'body', u'name': u'lotID'}
+    ])
+
+    response = self.app.post_json('/auctions/some_id/awards', {'data': {
+                                  'suppliers': [test_organization], 'bid_id': self.initial_bids[0]['id']}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+    response = self.app.get('/auctions/some_id/awards', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+    self.set_status('complete')
+
+    response = self.app.post_json('/auctions/{}/awards'.format(
+        self.auction_id), {'data': {'suppliers': [test_organization], 'status': 'pending', 'bid_id': self.initial_bids[0]['id']}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't create award in current (complete) auction status")
+
+
+def create_auction_award(self):
+    request_path = '/auctions/{}/awards'.format(self.auction_id)
+    response = self.app.post_json(request_path, {'data': {'suppliers': [test_organization], 'status': 'pending', 'bid_id': self.initial_bids[0]['id']}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    award = response.json['data']
+    self.assertEqual(award['suppliers'][0]['name'], test_organization['name'])
+    self.assertIn('id', award)
+    self.assertIn(award['id'], response.headers['Location'])
+
+    response = self.app.get(request_path)
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data'][-1], award)
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']['status'], u'active')
+
+    response = self.app.get('/auctions/{}'.format(self.auction_id))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']['status'], u'active.awarded')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "cancelled"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']['status'], u'cancelled')
+    self.assertIn('Location', response.headers)
+
+
+def patch_auction_award(self):
+    request_path = '/auctions/{}/awards'.format(self.auction_id)
+    response = self.app.post_json(request_path, {'data': {'suppliers': [test_organization], 'status': u'pending', 'bid_id': self.initial_bids[0]['id'], "value": {"amount": 500}}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    award = response.json['data']
+
+    #response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"value": {"amount": 600}}})
+    #self.assertEqual(response.status, '200 OK')
+    #self.assertEqual(response.content_type, 'application/json')
+    #self.assertEqual(response.json['data']["value"]["amount"], 600)
+
+    response = self.app.patch_json('/auctions/{}/awards/some_id'.format(self.auction_id), {"data": {"status": "unsuccessful"}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'award_id'}
+    ])
+
+    response = self.app.patch_json('/auctions/some_id/awards/some_id', {"data": {"status": "unsuccessful"}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"awardStatus": "unsuccessful"}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'], [
+        {"location": "body", "name": "awardStatus", "description": "Rogue field"}
+    ])
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "unsuccessful"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertIn('Location', response.headers)
+    new_award_location = response.headers['Location']
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "pending"}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't update award in current (unsuccessful) status")
+
+    response = self.app.get(request_path)
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(len(response.json['data']), 2)
+    self.assertIn(response.json['data'][1]['id'], new_award_location)
+    new_award = response.json['data'][-1]
+
+    response = self.app.patch_json('/auctions/{}/awards/{}?acc_token={}'.format(self.auction_id, new_award['id'], self.auction_token),
+                                   {"data": {"title": "title", "description": "description"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']['title'], 'title')
+    self.assertEqual(response.json['data']['description'], 'description')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, new_award['id']), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+
+    response = self.app.get(request_path)
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(len(response.json['data']), 2)
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, new_award['id']), {"data": {"status": "cancelled"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertIn('Location', response.headers)
+
+    response = self.app.get(request_path)
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(len(response.json['data']), 3)
+
+    self.set_status('complete')
+
+    response = self.app.get('/auctions/{}/awards/{}'.format(self.auction_id, award['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["value"]["amount"], 500)
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "unsuccessful"}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't update award in current (complete) auction status")
+
+
+def patch_auction_award_unsuccessful(self):
+    request_path = '/auctions/{}/awards'.format(self.auction_id)
+    response = self.app.post_json(request_path, {'data': {'suppliers': [test_organization], 'status': u'pending', 'bid_id': self.initial_bids[0]['id'], "value": {"amount": 500}}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    award = response.json['data']
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "unsuccessful"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertIn('Location', response.headers)
+    new_award_location = response.headers['Location']
+
+    response = self.app.patch_json(new_award_location[-82:], {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertNotIn('Location', response.headers)
+
+    response = self.app.get(request_path)
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(len(response.json['data']), 2)
+
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(self.auction_id, award['id']), {'data': {
+        'title': 'complaint title',
+        'description': 'complaint description',
+        'author': test_organization,
+        'status': 'claim'
+    }})
+    self.assertEqual(response.status, '201 Created')
+
+    response = self.app.post_json('{}/complaints'.format(new_award_location[-82:]), {'data': {
+        'title': 'complaint title',
+        'description': 'complaint description',
+        'author': test_organization
+    }})
+    self.assertEqual(response.status, '201 Created')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"status": "cancelled"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertIn('Location', response.headers)
+    new_award_location = response.headers['Location']
+
+    response = self.app.patch_json(new_award_location[-82:], {"data": {"status": "unsuccessful"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertIn('Location', response.headers)
+    new_award_location = response.headers['Location']
+
+    response = self.app.patch_json(new_award_location[-82:], {"data": {"status": "unsuccessful"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertNotIn('Location', response.headers)
+
+    response = self.app.get(request_path)
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(len(response.json['data']), 4)
+
+
+def get_auction_award(self):
+    response = self.app.post_json('/auctions/{}/awards'.format(
+        self.auction_id), {'data': {'suppliers': [test_organization], 'status': 'pending', 'bid_id': self.initial_bids[0]['id']}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    award = response.json['data']
+
+    response = self.app.get('/auctions/{}/awards/{}'.format(self.auction_id, award['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    award_data = response.json['data']
+    self.assertEqual(award_data, award)
+
+    response = self.app.get('/auctions/{}/awards/some_id'.format(self.auction_id), status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'award_id'}
+    ])
+
+    response = self.app.get('/auctions/some_id/awards/some_id', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+
+def patch_auction_award_Administrator_change(self):
+    response = self.app.post_json('/auctions/{}/awards'.format(
+        self.auction_id), {'data': {'suppliers': [test_organization], 'status': 'pending', 'bid_id': self.initial_bids[0]['id']}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    award = response.json['data']
+    complaintPeriod = award['complaintPeriod'][u'startDate']
+
+    authorization = self.app.authorization
+    self.app.authorization = ('Basic', ('administrator', ''))
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, award['id']), {"data": {"complaintPeriod": {"endDate": award['complaintPeriod'][u'startDate']}}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertIn("endDate", response.json['data']['complaintPeriod'])
+    self.assertEqual(response.json['data']['complaintPeriod']["endDate"], complaintPeriod)
+
+# AuctionAwardComplaintResourceTest
+
+
+def create_auction_award_complaint_invalid(self):
+    response = self.app.post_json('/auctions/some_id/awards/some_id/complaints', {
+                                  'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location': u'url', u'name': u'auction_id'}
+    ])
+
+    request_path = '/auctions/{}/awards/{}/complaints'.format(self.auction_id, self.award_id)
+
+    response = self.app.post(request_path, 'data', status=415)
+    self.assertEqual(response.status, '415 Unsupported Media Type')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description':
+            u"Content-Type header should be one of ['application/json']", u'location': u'header', u'name': u'Content-Type'}
+    ])
+
+    response = self.app.post(
+        request_path, 'data', content_type='application/json', status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Expecting value: line 1 column 1 (char 0)',
+            u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, 'data', status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Data not available',
+            u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(
+        request_path, {'not_data': {}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Data not available',
+            u'location': u'body', u'name': u'data'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': [u'This field is required.'], u'location': u'body', u'name': u'author'},
+        {u'description': [u'This field is required.'], u'location': u'body', u'name': u'title'},
+    ])
+
+    response = self.app.post_json(request_path, {'data': {
+                                  'invalid_field': 'invalid_value'}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Rogue field', u'location':
+            u'body', u'name': u'invalid_field'}
+    ])
+
+    response = self.app.post_json(request_path, {
+                                  'data': {'author': {'identifier': 'invalid_value'}}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': {u'identifier': [
+            u'Please use a mapping for this field or Identifier instance instead of unicode.']}, u'location': u'body', u'name': u'author'}
+    ])
+
+    response = self.app.post_json(request_path, {
+                                  'data': {'title': 'complaint title', 'description': 'complaint description', 'author': {'identifier': {'id': 0}}}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': {u'contactPoint': [u'This field is required.'], u'identifier': {u'scheme': [u'This field is required.']}, u'name': [u'This field is required.'], u'address': [u'This field is required.']}, u'location': u'body', u'name': u'author'}
+    ])
+
+    response = self.app.post_json(request_path, {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': {
+        'name': 'name', 'identifier': {'uri': 'invalid_value'}}}}, status=422)
+    self.assertEqual(response.status, '422 Unprocessable Entity')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': {u'contactPoint': [u'This field is required.'], u'identifier': {u'scheme': [u'This field is required.'], u'id': [u'This field is required.'], u'uri': [u'Not a well formed URL.']}, u'address': [u'This field is required.']}, u'location': u'body', u'name': u'author'}
+    ])
+
+
+def create_auction_award_complaint(self):
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization, 'status': 'claim'}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    complaint = response.json['data']
+    self.assertEqual(complaint['author']['name'], test_organization['name'])
+    self.assertIn('id', complaint)
+    self.assertIn(complaint['id'], response.headers['Location'])
+
+    self.set_status('active.awarded')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], self.auction_token), {"data": {
+        "status": "answered",
+        "resolutionType": "invalid",
+        "resolution": "spam 100% " * 3
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "answered")
+    self.assertEqual(response.json['data']["resolutionType"], "invalid")
+    self.assertEqual(response.json['data']["resolution"], "spam 100% " * 3)
+
+    response = self.app.get('/auctions/{}'.format(self.auction_id))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], 'active.awarded')
+
+    self.set_status('unsuccessful')
+
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't add complaint in current (unsuccessful) auction status")
+
+
+def patch_auction_award_complaint(self):
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    complaint = response.json['data']
+    owner_token = response.json['access']['token']
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], self.auction_token), {"data": {
+        "status": "cancelled",
+        "cancellationReason": "reason"
+    }}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Forbidden")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, self.award_id, complaint['id']), {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "active")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "title": "claim title",
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']["title"], "claim title")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "status": "claim",
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.json['data']["status"], "claim")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], self.auction_token), {"data": {
+        "resolution": "changing rules"
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["resolution"], "changing rules")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], self.auction_token), {"data": {
+        "status": "answered",
+        "resolutionType": "resolved",
+        "resolution": "resolution text " * 2
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "answered")
+    self.assertEqual(response.json['data']["resolutionType"], "resolved")
+    self.assertEqual(response.json['data']["resolution"], "resolution text " * 2)
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "satisfied": False
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["satisfied"], False)
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "status": "resolved"
+    }}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't update complaint")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "status": "pending"
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "pending")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "status": "cancelled",
+        "cancellationReason": "reason"
+    }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "cancelled")
+    self.assertEqual(response.json['data']["cancellationReason"], "reason")
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/some_id'.format(self.auction_id, self.award_id), {"data": {"status": "resolved", "resolution": "resolution text"}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'complaint_id'}
+    ])
+
+    response = self.app.patch_json('/auctions/some_id/awards/some_id/complaints/some_id', {"data": {"status": "resolved", "resolution": "resolution text"}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "status": "cancelled",
+        "cancellationReason": "reason"
+    }}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't update complaint in current (cancelled) status")
+
+    response = self.app.patch_json('/auctions/{}/awards/some_id/complaints/some_id'.format(self.auction_id), {"data": {"status": "resolved", "resolution": "resolution text"}}, status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'award_id'}
+    ])
+
+    response = self.app.get('/auctions/{}/awards/{}/complaints/{}'.format(self.auction_id, self.award_id, complaint['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "cancelled")
+    self.assertEqual(response.json['data']["cancellationReason"], "reason")
+    self.assertEqual(response.json['data']["resolutionType"], "resolved")
+    self.assertEqual(response.json['data']["resolution"], "resolution text " * 2)
+
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    complaint = response.json['data']
+    owner_token = response.json['access']['token']
+
+    self.set_status('complete')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+        "status": "claim",
+    }}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can't update complaint in current (complete) auction status")
+
+
+def review_auction_award_complaint(self):
+    complaints = []
+    for i in range(3):
+        response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(self.auction_id, self.award_id), {'data': {
+            'title': 'complaint title',
+            'description': 'complaint description',
+            'author': test_organization,
+            'status': 'claim'
+        }})
+        self.assertEqual(response.status, '201 Created')
+        self.assertEqual(response.content_type, 'application/json')
+        complaint = response.json['data']
+        owner_token = response.json['access']['token']
+        complaints.append(complaint)
+
+        response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], self.auction_token), {"data": {
+            "status": "answered",
+            "resolutionType": "resolved",
+            "resolution": "resolution text " * 2
+        }})
+        self.assertEqual(response.status, '200 OK')
+        self.assertEqual(response.content_type, 'application/json')
+        self.assertEqual(response.json['data']["status"], "answered")
+        self.assertEqual(response.json['data']["resolutionType"], "resolved")
+        self.assertEqual(response.json['data']["resolution"], "resolution text " * 2)
+
+        response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}?acc_token={}'.format(self.auction_id, self.award_id, complaint['id'], owner_token), {"data": {
+            "satisfied": False,
+            "status": "pending"
+        }})
+        self.assertEqual(response.status, '200 OK')
+        self.assertEqual(response.content_type, 'application/json')
+        self.assertEqual(response.json['data']["status"], "pending")
+
+    self.app.authorization = ('Basic', ('reviewer', ''))
+    for complaint, status in zip(complaints, ['invalid', 'resolved', 'declined']):
+        response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}'.format(self.auction_id, self.award_id, complaint['id']), {"data": {
+            "decision": '{} complaint'.format(status)
+        }})
+        self.assertEqual(response.status, '200 OK')
+        self.assertEqual(response.content_type, 'application/json')
+        self.assertEqual(response.json['data']["decision"], '{} complaint'.format(status))
+
+        response = self.app.patch_json('/auctions/{}/awards/{}/complaints/{}'.format(self.auction_id, self.award_id, complaint['id']), {"data": {
+            "status": status
+        }})
+        self.assertEqual(response.status, '200 OK')
+        self.assertEqual(response.content_type, 'application/json')
+        self.assertEqual(response.json['data']["status"], status)
+
+
+def get_auction_award_complaint(self):
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    complaint = response.json['data']
+
+    response = self.app.get('/auctions/{}/awards/{}/complaints/{}'.format(self.auction_id, self.award_id, complaint['id']))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data'], complaint)
+
+    response = self.app.get('/auctions/{}/awards/{}/complaints/some_id'.format(self.auction_id, self.award_id), status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'complaint_id'}
+    ])
+
+    response = self.app.get('/auctions/some_id/awards/some_id/complaints/some_id', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+
+def get_auction_award_complaints(self):
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}})
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    complaint = response.json['data']
+
+    response = self.app.get('/auctions/{}/awards/{}/complaints'.format(self.auction_id, self.award_id))
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data'][0], complaint)
+
+    response = self.app.get('/auctions/some_id/awards/some_id/complaints', status=404)
+    self.assertEqual(response.status, '404 Not Found')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['status'], 'error')
+    self.assertEqual(response.json['errors'], [
+        {u'description': u'Not Found', u'location':
+            u'url', u'name': u'auction_id'}
+    ])
+
+    auction = self.db.get(self.auction_id)
+    for i in auction.get('awards', []):
+        i['complaintPeriod']['endDate'] = i['complaintPeriod']['startDate']
+    self.db.save(auction)
+
+    response = self.app.post_json('/auctions/{}/awards/{}/complaints'.format(
+        self.auction_id, self.award_id), {'data': {'title': 'complaint title', 'description': 'complaint description', 'author': test_organization}}, status=403)
+    self.assertEqual(response.status, '403 Forbidden')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['errors'][0]["description"], "Can add complaint only in complaintPeriod")
