+++ openprocurement/auctions/core/tests/award.py
@@ -10,19 +10,6 @@ from openprocurement.auctions.core.tests.blanks.award_blanks import (
     # Auction2LotAwardResourceTest
     create_auction_award_2_lots,
     patch_auction_award_2_lots,
-    invalid_patch_auction_award,
-    patch_auction_award,
-    patch_auction_award_admin,
-    complate_auction_with_second_award1,
-    complate_auction_with_second_award2,
-    complate_auction_with_second_award3,
-    unsuccessful_auction1,
-    unsuccessful_auction2,
-    unsuccessful_auction3,
-    unsuccessful_auction4,
-    unsuccessful_auction5,
-    get_auction_awards,
-    successful_second_auction_award,
     not_found, create_auction_award_document,
     put_auction_award_document,
     patch_auction_award_document,
@@ -57,22 +44,6 @@ class AuctionLotAwardResourceTestMixin(object):
 class Auction2LotAwardResourceTestMixin(object):
     test_create_auction_award_2_lots = snitch(create_auction_award_2_lots)
     test_patch_auction_award_2_lots = snitch(patch_auction_award_2_lots)
-
-
-class AuctionAwardProcessTestMixin(object):
-    test_invalid_patch_auction_award = snitch(invalid_patch_auction_award)
-    test_patch_auction_award = snitch(patch_auction_award)
-    test_patch_auction_award_admin = snitch(patch_auction_award_admin)
-    test_complate_auction_with_second_award1 = snitch(complate_auction_with_second_award1)
-    test_complate_auction_with_second_award2 = snitch(complate_auction_with_second_award2)
-    test_complate_auction_with_second_award3 = snitch(complate_auction_with_second_award3)
-    test_successful_second_auction_award = snitch(successful_second_auction_award)
-    test_unsuccessful_auction1 = snitch(unsuccessful_auction1)
-    test_unsuccessful_auction2 = snitch(unsuccessful_auction2)
-    test_unsuccessful_auction3 = snitch(unsuccessful_auction3)
-    test_unsuccessful_auction4 = snitch(unsuccessful_auction4)
-    test_unsuccessful_auction5 = snitch(unsuccessful_auction5)
-    test_get_auction_awards = snitch(get_auction_awards)
 
 
 class AuctionAwardDocumentResourceTestMixin(object):
